Spotify-Downloader/requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-celery-results==2.5.1
django-db-geventpool==4.0.7
gevent==24.11.1
greenlet==3.1.1
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
pycparser==2.22
pyreqs==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytubefix==8.8.2
requests==2.32.3
setuptools==75.6.0
sh==2.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
